[WIP] fixed duplicate checking, dry run; discovered both addonly and forcekeep don't work
- fixed duplicate checking,
- made  copying of  bams under control of dry run
- discovered both addonly and forcekeep don't work:
-- forcekeep  functionality doesn't work due to some of the details of implementing the duplicate checking
-- addonly doesn't work because notfound variable is created but never used to determine which  samples  to add to Terra

+++ src/CCLE_postp_function.py
@@ -14,6 +14,7 @@ import numpy as np
 from taigapy import TaigaClient
 tc = TaigaClient()
 import os
+import re # for re.split
 
 CHROMLIST = ['chr1', 'chr2', 'chr3', 'chr4', 'chr5', 'chr6', 'chr7', 'chr8',
              'chr9', 'chr10', 'chr11', 'chr12', 'chr13', 'chr14', 'chr15',
@@ -36,7 +37,8 @@ extract_defaults = {
     'hs_on_target_bases': 'hs_on_target_bases',
     'total_reads': 'total_reads',
     'release_date': 'release_date',
-    'mean_depth': 'mean_depth'
+    'mean_depth': 'mean_depth',
+    'ref_bams':'bam'
 }
 
 
@@ -46,11 +48,11 @@ def read(filename):
   f.close()
   return x
 
-def get_all_bai_size(folder):
-  # get ls of all bai files in the folder
-  samples = os.popen('gsutil -m ls -al ' + folder + '**.bai').read().split('\n')
+def get_all_sizes(folder, suffix):
+  # get ls of all files in the folder with the specified suffix
+  samples = os.popen('gsutil -m ls -al ' + folder + '**.' + suffix).read().split('\n')
   # compute size filepath
-  sizes = {'gs://' + val.split('gs://')[1].split('#')[0]: int(val.split("2019-")[0]) for val in samples[:-2]}
+  sizes = {'gs://' + val.split('gs://')[1].split('#')[0]: int(re.split("\d{4}-\d{2}-\d{2}", val)[0]) for val in samples[:-2]}
   names = {}
   for k, val in sizes.items():
     if val in names:
@@ -59,12 +61,28 @@ def get_all_bai_size(folder):
       names[val] = [k]
   return names
 
+def get_GS_file_sizes(file_list):
+  # get size of each file in the input list of files (assumed to be in google storage)
+  for file in file_list:
+    samples = os.popen('gsutil -m ls -al ' + file).read().split('\n')
+    # compute size filepath
+    sizes = {'gs://' + val.split('gs://')[1].split('#')[0]: int(re.split("\d{4}-\d{2}-\d{2}", val)[0]) for val in samples[:-2]}
+    names = {}
+    for k, val in sizes.items():
+      if val in names:
+        names[val].append(k)
+      else:
+        names[val] = [k]
+  return names
+
 def createDatasetWithNewCellLines(wto, samplesetname,
                                   wmfroms, sources,
-                                  forcekeep=[], addonly=[], match='ACH', other_to_add=[], extract=extract_defaults,
+                                  forcekeep=[], addonly=[], match=['ACH'], other_to_add=[], extract=extract_defaults,
                                   participantslicepos=10, accept_unknowntypes=False,
                                   gsfolderto=None, dry_run = False):
   """
+  ## check: will need to change slicepos based on match;
+  ## may have to take diff approach (regex?) for CCLF samples since CCLF sample IDs are not of consistent length
   wto: dalmatian.workspacemanager the workspace where you want to create the tsvs
   samplesetname: String the name of the sample set created by this new set of samples updated
   wfrom1: dalmatian.workspacemanager the workspace where the Samples to add are stored
@@ -74,7 +92,7 @@ def createDatasetWithNewCellLines(wto, samplesetname,
   forcekeep: list of sample id that you want to keep even if already in the previous workspace (will
   cause an overwrite) NOTE: currently this functionality is broken due to the implementation of checking for duplicates
   addonly: list of sample id that you only want to add
-  match: substring that has to be matched against the id of the samples to add them
+  match: list of substring(s) that has to be matched against the id of the samples to add them
   other_to_add:
   extract: if you want to specify what values should refer to what
     dict{    'name':
@@ -86,11 +104,12 @@ def createDatasetWithNewCellLines(wto, samplesetname,
 
   """
   refsamples = wto.get_samples()
-  refids = refsamples['participant'].tolist()
+  refids_full = refsamples['participant'].tolist()
   # do NOT make refids a set; we use the num of occurences as way to determine what number to add to the sample id
-  refids = [val[val.index('ACH'):] for val in refids if 'ACH' in val]
-  print("Getting sizes of all bai files we have to we can check for duplicates later...")
-  bai_sizes = get_all_bai_size(folder = gsfolderto)
+  # filter refids to only include those that include the strings in the 'match' argument
+  refids = []
+  for match_substring in match:
+      refids += [val[val.index(match_substring):] for val in refids_full if match_substring in val]
 
   print("Getting sample infos...")
   if type(sources) is str:
@@ -101,36 +120,50 @@ def createDatasetWithNewCellLines(wto, samplesetname,
   for source, wmfrom in zip(sources, wmfroms):
     samples = wmfrom.get_samples().replace(np.nan, '', regex=True).reset_index()
     # keep samples that contain the match requirement (e.g. ACH for DepMap IDs)
-    samples = samples[samples[extract['id']].str.contains(match)]
+    samples = samples[samples[extract['id']].str.contains('|'.join(match))]
+    print("\nThe shape of the sample tsv from " + str(wmfrom) + ": " + str(samples.shape))
 
-    # remove true duplicates from consideration
-    print("Identifying any true duplicates by checking file sizes (this happens for each data source)...")
+    ## remove true duplicates from consideration
+    print("Identifying any true duplicates by checking file sizes (this runs for each data source)...")
+    print("This step can take a while as we need to use gsutil to check the size of each potential duplicate...")
     dups_to_remove = []
     for k, rowinfo in samples.iterrows():
-      bai_filepath = rowinfo[extract['bai']]
-      sample_to_check = os.popen('gsutil -m ls -al ' + bai_filepath).read().split('\n')
-      # sample_size is a dict with one key only
-      sample_size = {'gs://' + val.split('gs://')[1].split('#')[0]: int(val.split("2019-")[0]) for val in sample_to_check[:-2]}
-      for k, val in sample_size.items():
-        # if seen this size bai file before, it's a duplicate
-        if val in bai_sizes:
-          dups_to_remove += [rowinfo[extract['id']]]
-    # check: currently, below lines prevent forcekeep from working on true duplicates (aka same size bai file). Is this a problem?
-    samples = samples[(samples[extract['id']].str.slice(0, participantslicepos).isin(dups_to_remove))]
+      # if this sample is in Terra already, need to check for duplicates
+      if rowinfo[extract['id']][0:participantslicepos] in refids:
+        filepath = rowinfo[extract['bam']]
+        sample_to_check = os.popen('gsutil -m ls -al ' + filepath).read().split('\n')
+        # better regex for yyyy-mm-dd ([12]\d{3}-(0[1-9]|1[0-2])-(0[1-9]|[12]\d|3[01]))
+        sample_size = {'gs://' + val.split('gs://')[1].split('#')[0]: int(re.split("\d{4}-\d{2}-\d{2}", val)[0]) for val in sample_to_check[:-2]}
+        for file, val in sample_size.items():
+          # get list of the pre-existing entries for the same participant
+          same_participant = refsamples[refsamples['participant'] == rowinfo[extract['id']][0:participantslicepos]]
+          same_participant_bams = []
+          for samp, entry in same_participant.iterrows():
+            if entry[extract['ref_bams']] != 'NA':
+              same_participant_bams += [entry[extract['ref_bams']]]
+          # get sizes of these pre-existing bams
+          same_participant_bams_sizes = get_GS_file_sizes(same_participant_bams)
+          # if new sample size is in this dict of sizes, then the sample is a duplicate
+          if val in same_participant_bams_sizes:
+            dups_to_remove += [rowinfo[extract['id']]]
+          # print("The new sample: " + str(file))
+          # print("The pre-existing sample of the same size: " + str(same_participant_bams_sizes[val]))
 
-    # add number to sample ID so runs of same participant have unique sample ID
+    # check: currently, below lines prevent forcekeep from working on true duplicates (aka same size  file). Need to think about how to bring back the forcekeep functionality
+    print("Dups from " + str(wmfrom) + " has len " + str(len(dups_to_remove)) + ":\n " + str(dups_to_remove))
+    print("Len of samples before removal: " + str(len(samples)))
+    # remove the duplicates from consideration
+    samples = samples[(~samples[extract['id']].isin(dups_to_remove))]
+    samples.reset_index(drop=True, inplace=True)
+    print("Len of samples after removal: " + str(len(samples)))
+
+    ## add number to sample ID so runs of same participant have unique sample ID
     new_samples = samples[extract['id']].str.slice(0, participantslicepos)
     for iSample in range(len(new_samples)):
       new_sample_id = new_samples.values[iSample]
       num_to_add = refids.count(new_sample_id) + 1
       # update the sample dataframe with the new sample ID
-      # check: fix the column name; should be in same order I believe
       samples[extract['id']][iSample] = new_sample_id + '_' + str(num_to_add)
-
-    ## no longer using this stuff, but need to think about how to bring back the forcekeep functionality:
-    # samples[samples[extract['id']].str.slice(0, participantslicepos).isin(refids)]
-    # [(~samples[extract['id']].str.slice(0, participantslicepos).isin(refids)) |
-    # (samples[extract['id']].isin(forcekeep))]
 
     if len(addonly) > 0:
       samples = samples[samples[extract['id']].isin(addonly)]
@@ -140,21 +173,24 @@ def createDatasetWithNewCellLines(wto, samplesetname,
       if not accept_unknowntypes:
         samples = samples[samples['sample_type'].isin(['Tumor'])]
     sampless = pd.concat([sampless, samples], sort=False)
-  notfound = set(addonly) - set(sampless.index.tolist())
-  if len(notfound) > 0:
-    # samples not found in the wto workspace so will be adding them now
-    print('we did not find:' + str(notfound))
+  # ## currently, we don't do anything with notfound. Therefore addonly functionality is broken.
+  # notfound = set(addonly) - set(sampless.index.tolist())
+  # if len(notfound) > 0:
+  #   # samples not found in the wto workspace so will be adding them now
+  #   print('we did not find:' + str(notfound))
+
   sample_ids = []
   # to do the download to the new dataspace
-  if gsfolderto is not None:
-    for i, val in sampless.iterrows():
-      res = os.system('gsutil cp ' + val[extract['bam']] + ' ' + val[extract['bai']] + ' ' + gsfolderto)
-      if res == 256:
-        sampless.drop(i)
-        print('we got sample ' + str(i) + ' that had nonexistant bam path')
-    sampless[extract['bam']] = [gsfolderto + a.split('/')[-1] for a in sampless[extract['bam']]]
-    sampless[extract['bai']] = [gsfolderto + a.split('/')[-1] for a in sampless[extract['bai']]]
-  print(sampless)
+  if not dry_run:
+    if gsfolderto is not None:
+      for i, val in sampless.iterrows():
+        res = os.system('gsutil cp ' + val[extract['bam']] + ' ' + val[extract['bai']] + ' ' + gsfolderto)
+        if res == 256:
+          sampless.drop(i)
+          print('we got sample ' + str(i) + ' that had nonexistant bam path')
+      sampless[extract['bam']] = [gsfolderto + a.split('/')[-1] for a in sampless[extract['bam']]]
+      sampless[extract['bai']] = [gsfolderto + a.split('/')[-1] for a in sampless[extract['bai']]]
+    print(sampless)
   if len(sampless) == 0:
     raise Exception("no new samples in this matrix")
   for ind, val in sampless.iterrows():
@@ -175,7 +211,7 @@ def createDatasetWithNewCellLines(wto, samplesetname,
       wto.upload_samples(refsamples)
       print("creating a sample set")
       wto.update_sample_set(sample_set_id=samplesetname, sample_ids=sample_ids)
-  return sample_ids, refsamples['WES_bam'].values
+  return sample_ids, refsamples['WES_bam'].values, refsamples
 
 #####################
 # VALIDATION
